Update example kanban with new css
