+++ .env
@@ -1,3 +1,5 @@
 NEXT_PUBLIC_API_ADDR='http://localhost:8080'
 NEXT_PUBLIC_DASHBOARD_PATH='/dashboard'
 NEXT_PUBLIC_JS_SDK_URL='https://cdnjs.cloudflare.com/ajax/libs/yorkie-js-sdk/0.2.16/yorkie-js-sdk.js'
+
+VITE_YORKIE_API_KEY='cedaovjuioqlk4pjqn6g'
