+++ .env.production
@@ -1,2 +1,4 @@
 NEXT_PUBLIC_BASE_PATH='/homepage'
 NEXT_PUBLIC_API_ADDR='https://api.yorkie.dev'
+
+VITE_YORKIE_API_KEY='cedaovjuioqlk4pjqn6g'
