--- components/BasicExampleCodes/index.ts
@@ -1,14 +1,0 @@
-interface File {
-  type: string;
-  name: string;
-  content: string;
-}
-interface NestedFolder {
-  [key: string]: NestedFolder | File;
-}
-
-export interface ProjectStructure {
-  [key: string]: NestedFolder | File;
-}
-
-export * from './kanban';
